Don't modify game clocks on game end to avoid confusion
While it is useful to get remaining clock times when the game is
finished, it is confusing when the game is just finished and ends with a
timeout.

Instead we'll only show remaining clock times when accessing a game from
the archived games list, because here there is no possible confusion.

Fixes: #440