Fix plural handling for zero quantity
The removed code in mjs assumed that in all languages
`0` is followed by singular but that's not true e.g. for German and English.

I tested with French after removing the workaround and
updating localizations: `intl.Intl.pluralLogic` seems to handle
the plural logic correctly, choosing singular after 0 for French,
but plural for German and English.

Probably at the time that this code was added, `intl.Intl.pluralLogic`
was not used yet...?