fix: empty space in "Watch" tab in landscape orientation
The previous code would hide the _BroadcastWidget widget by returning
SizedBox.shrink(), but that would still create an empty grid space in
landscape mode.

To fix this, factor out the widget list into a getter (to avoid code
duplication) and comment out the BroadcastWidget in that list for now.

Fixes #829