Ensure popover background syncs with system theme changes
Closes #1881

Note that this bug applies to all places where popover is used.
This is not really a proper fix because the arrow tip of popover won't
change and will show as white in a dark background. But at least the
text is visible. There is no obvious way to fix this and it is only an
edge case, so there is no urge to take time to do better here.