Change the computer analysis toggle behavior
Before it acted on both local and server analysis. With this commit it
now acts only on server analysis.
The reason for this is to allow people who don't want to be spoiled by
automatic server analysis to disable completely the feature, and still
allow them to analyse the game themselves, and use the local engine if
they want to.