fix: don't try to account for alternative castling notation in annotations (#2255)
* fix: don't try to account for alternative castling notation in annotations

This leads annotations being displayed on the wrong square e.g. when
moving the rook from e1 to h1 (as its incorecctly interpreted as a
castling move).

Luckily, we already convert alternative castling notation when constructing the
node tree, so instead of extending the logic in the AnalysisBoard, we
can just get rid of this logic entirely.

Fixes #2231

* add regression test