fix regression caused by #2255 (#2272)
I wrongly assumed that addMoveAt() already converts castling notation
for us, but that function is only called when new nodes are added to the
move tree in an active game, but not when parsing a study chapter.

This is why #2255 fixed the bug of a mve being incorrectly interpreted
as alt-castling, but it now broke the annotations if the move actually
WAS a castling move.

Luckily, while in addMoveAt() there are some heuristics because we only
have the UCI move, not the SAN move, in the AnalysisBoard we *do* have
the SAN move, so we can just check for O-O/O-O-O to avoid false
positives.

Resolves #2231